requirements.txt
llama-index
llama-hub
openai
python-dotenv